ExoPlayerListener is a nested class
