+++ src/org/thoughtcrime/securesms/video/VideoPlayer.java
@@ -16,29 +16,35 @@
  */
 package org.thoughtcrime.securesms.video;
 
-import android.app.Activity;
 import android.content.Context;
 import android.os.Build;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.util.AttributeSet;
 import android.util.Log;
+import android.view.Window;
+import android.view.WindowManager;
 import android.widget.FrameLayout;
 import android.widget.MediaController;
 import android.widget.Toast;
 import android.widget.VideoView;
 
 import com.google.android.exoplayer2.DefaultLoadControl;
+import com.google.android.exoplayer2.ExoPlaybackException;
+import com.google.android.exoplayer2.ExoPlayer;
 import com.google.android.exoplayer2.ExoPlayerFactory;
 import com.google.android.exoplayer2.LoadControl;
 import com.google.android.exoplayer2.SimpleExoPlayer;
+import com.google.android.exoplayer2.Timeline;
 import com.google.android.exoplayer2.extractor.DefaultExtractorsFactory;
 import com.google.android.exoplayer2.extractor.ExtractorsFactory;
 import com.google.android.exoplayer2.source.ExtractorMediaSource;
 import com.google.android.exoplayer2.source.MediaSource;
+import com.google.android.exoplayer2.source.TrackGroupArray;
 import com.google.android.exoplayer2.trackselection.AdaptiveTrackSelection;
 import com.google.android.exoplayer2.trackselection.DefaultTrackSelector;
 import com.google.android.exoplayer2.trackselection.TrackSelection;
+import com.google.android.exoplayer2.trackselection.TrackSelectionArray;
 import com.google.android.exoplayer2.trackselection.TrackSelector;
 import com.google.android.exoplayer2.ui.SimpleExoPlayerView;
 import com.google.android.exoplayer2.upstream.BandwidthMeter;
@@ -52,7 +58,6 @@ import org.thoughtcrime.securesms.mms.PartAuthority;
 import org.thoughtcrime.securesms.mms.VideoSlide;
 import org.thoughtcrime.securesms.util.ViewUtil;
 import org.thoughtcrime.securesms.video.exo.AttachmentDataSourceFactory;
-import org.thoughtcrime.securesms.video.exo.ExoPlayerListener;
 
 import java.io.IOException;
 
@@ -65,6 +70,7 @@ public class VideoPlayer extends FrameLayout {
 
   @Nullable private       SimpleExoPlayer     exoPlayer;
   @Nullable private       AttachmentServer    attachmentServer;
+  @Nullable private       Window              window;
 
   public VideoPlayer(Context context) {
     this(context, null);
@@ -106,6 +112,10 @@ public class VideoPlayer extends FrameLayout {
     }
   }
 
+  public void setWindow(Window window) {
+    this.window = window;
+  }
+
   private void setExoViewSource(@NonNull MasterSecret masterSecret, @NonNull VideoSlide videoSource)
       throws IOException
   {
@@ -115,7 +125,7 @@ public class VideoPlayer extends FrameLayout {
     LoadControl            loadControl                = new DefaultLoadControl();
 
     exoPlayer = ExoPlayerFactory.newSimpleInstance(getContext(), trackSelector, loadControl);
-    exoPlayer.addListener(new ExoPlayerListener(((Activity) this.getContext()).getWindow()));
+    exoPlayer.addListener(new ExoPlayerListener(window));
     exoView.setPlayer(exoPlayer);
 
     DefaultDataSourceFactory    defaultDataSourceFactory    = new DefaultDataSourceFactory(getContext(), "GenericUserAgent", null);
@@ -158,5 +168,48 @@ public class VideoPlayer extends FrameLayout {
     mediaController.setMediaPlayer(videoView);
 
     videoView.setMediaController(mediaController);
+  }
+
+  private class ExoPlayerListener implements ExoPlayer.EventListener {
+    private final Window window;
+
+    ExoPlayerListener(Window window) {
+      this.window = window;
+    }
+
+    @Override
+    public void onPlayerStateChanged(boolean playWhenReady, int playbackState) {
+      switch(playbackState) {
+        case ExoPlayer.STATE_IDLE:
+        case ExoPlayer.STATE_BUFFERING:
+        case ExoPlayer.STATE_ENDED:
+          window.clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
+          break;
+        case ExoPlayer.STATE_READY:
+          if (playWhenReady) {
+            window.addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
+          } else {
+            window.clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON);
+          }
+          break;
+        default:
+          break;
+      }
+    }
+
+    @Override
+    public void onTimelineChanged(Timeline timeline, Object manifest) { }
+
+    @Override
+    public void onTracksChanged(TrackGroupArray trackGroups, TrackSelectionArray trackSelections) { }
+
+    @Override
+    public void onLoadingChanged(boolean isLoading) { }
+
+    @Override
+    public void onPlayerError(ExoPlaybackException error) { }
+
+    @Override
+    public void onPositionDiscontinuity() { }
   }
 }
